plugins/AmapMCP/requirements.txt
aiohttp>=3.8.0
sseclient-py>=1.7.2
uuid>=1.30 